requirements.txt
a2a-sdk[http-server]>=0.3.0
click>=8.1.8
python-dotenv>=1.1.0
wikipedia>=1.4.0
openai>=2.11.0
uvicorn>=0.38.0